requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
